backend/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.15
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-pinecone>=0.2.0
langchain-deepseek>=0.1.0
langchain-text-splitters>=0.3.0   
langgraph>=0.2.20
llama-parse>=0.5.0
python-dotenv>=1.0.1
ddgs>=5.0.0
pinecone>=5.0.1
beautifulsoup4>=4.12.3
sentence-transformers>=3.0.1
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
redis>=5.2.0
langgraph-checkpoint-redis>=0.3.2
nest_asyncio>=1.5.0
requests>=2.31.0
langgraph-checkpoint-redis>=0.3.2